More runtime errors fixed
unsafe casts produce runtime errors, this is fixed now.

+++ src/io/colyseus/serializer/schema/Schema.hx
@@ -555,9 +555,10 @@ class Schema {
           var items = cast(valueRef.items, Array<Dynamic>);
 
           for (i in newLength...valueRef.items.length) {
-            var itemRemoved = items[i];
+			var itemRemoved = items[i];
+			trace('=========',itemRemoved);
 
-            if (isSchemaType) {
+            if (isSchemaType && itemRemoved.onRemove != null) {
               itemRemoved.onRemove();
             }
 
@@ -586,7 +587,7 @@ class Schema {
             if (isNew) {
               item = Type.createInstance(type, []);
             } else if (indexChangedFrom != -1) {
-              item = cast(valueRef, ArraySchema<Schema>).items[indexChangedFrom];
+              item = (cast valueRef).items[indexChangedFrom];
             } else {
               item = (cast valueRef).items[newIndex];
             }
